+++ kanbanflow/app/models/task.rb
@@ -1,4 +1,6 @@
 class Task < ApplicationRecord
+  include ActionView::RecordIdentifier
+
   belongs_to :project
 
   enum :status, { todo: 0, in_progress: 1, done: 2 }